Ensure craftable and displayed texture blocks match
Contents of item tags and the cooresponding block tags are not always the same, so use the item tag and just filter to contain blocks. Done since the recipe parser uses item tags.
Also make JEI blank non-block items in the item tag instead of filtering, to ensure the ingredient matches the output size for the animation