Add cauldron freezing in properly and combine cauldron block
Cauldron now stores temperature in the TE, so no need for a separate boiling block as that was causing issues with content updates (best I could do was missing texture showed for a half second on boiling update)
Freezing is defined by a cauldron having ice on opposite sides or being in a cold biome. Boiling now also happens in the nether.