Port code to 1.20
Strip out IHidable as its no longer needed with the new creative tab system. Debatably it is still useful for JEI, but that only really was needed for server configs; for non-server configs I don't mind that a pack reload is needed to resync as you need that for recipe updates anyways.
Fix config pack adding resources too late; was a consequence of a bug fix that was not fully tested on 1.19, so will get a backport.
Remove smithing recipe for custom bottles, they made more sense before I came up with the brewing recipe